feat: Update JWT authentication configuration
This commit updates the JWT authentication configuration by adding a new scheme called `SimpleJWTCookieScheme` in the `aura.core.schema` module. This scheme extends the `SimpleJWTScheme` and is used for JWT cookie authentication. The `target_class` is set to `aura.core.authentication.JWTCookieAuthentication` and the scheme is marked as optional. The name of the scheme is set to `jwtHeaderAuth` and `jwtCookieAuth`. This change enhances the authentication capabilities of the application.