feat: Add RestAuthJWTSerializer and CookieTokenRefreshSerializerExtension
This commit adds two new classes to the `aura.core.schema` module: `RestAuthJWTSerializer` and `CookieTokenRefreshSerializerExtension`.

`RestAuthJWTSerializer` is an extension of the `OpenApiSerializerExtension` and is used to map the `JWTSerializer` class from the `aura.users.api.serializers` module. It includes a `map_serializer` method that returns a modified version of the `JWTSerializer` class with an additional `user` field using the `UserDetailsSerializer`.

`CookieTokenRefreshSerializerExtension` is an extension of the `TokenRefreshSerializerExtension` and is used to map the `CookieTokenRefreshSerializer` class from the `aura.core.authentication` module. It is marked as optional and returns the `TokenRefresh` name.

These changes enhance the authentication and token refresh functionality of the application.